Change var SessionTimeRemain from float to double

+++ src/main/java/com/joffrey/irsdkjava/library/info/model/RaceInfo.java
@@ -14,11 +14,11 @@ public class RaceInfo {
     private String eventType;
 
     // Info on remaining
-    private float remainingTime;  // SessionTimeRemain
-    private int   remainingLaps;  // SessionLapsRemain
-    private float fuelLevel;      // FuelLevel
-    private float fuelLevelPct;   // FuelLevelPct
-    private float fuelUsePerHour; // FuelUsePerHour
+    private double remainingTime;  // SessionTimeRemain
+    private int    remainingLaps;  // SessionLapsRemain
+    private float  fuelLevel;      // FuelLevel
+    private float  fuelLevelPct;   // FuelLevelPct
+    private float  fuelUsePerHour; // FuelUsePerHour
 
     // Rating
     private int averageSOF;
